[#34] Refactor: Redis 설정 및 스크립트 core 모듈로 이동
- PostVoteRedisRepository 추가
