+++ devtribe-feed-consumer/src/main/java/com/devtribe/task/PostUnvoteTask.java
@@ -1,26 +1,22 @@
 package com.devtribe.task;
 
+import com.devtribe.domain.vote.dao.PostVoteRedisRepository;
 import com.devtribe.event.PostVoteEvent;
-import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.script.RedisScript;
 import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
-public class PostUnvoteTask extends AbstractVoteTask {
+public class PostUnvoteTask {
 
-    public PostUnvoteTask(
-        RedisScript<List> postUnvoteScript,
-        RedisTemplate<String, String> redisTemplate,
-        RedisScript<List> postVoteScript
-    ) {
-        super(postUnvoteScript, redisTemplate, postVoteScript);
+    private final PostVoteRedisRepository postVoteRedisRepository;
+
+    public PostUnvoteTask(PostVoteRedisRepository postVoteRedisRepository) {
+        this.postVoteRedisRepository = postVoteRedisRepository;
     }
 
     public void processEvent(PostVoteEvent event) {
-        unvote(
+        postVoteRedisRepository.unvote(
             event.getPostId(),
             event.getUserId()
         );
