+++ devtribe-feed-core/src/main/java/com/devtribe/domain/vote/dao/PostVoteRedisRepository.java
@@ -1,11 +1,13 @@
-package com.devtribe.task;
+package com.devtribe.domain.vote.dao;
 
 import com.devtribe.domain.vote.entity.VoteType;
 import java.util.List;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.script.RedisScript;
+import org.springframework.stereotype.Repository;
 
-public abstract class AbstractVoteTask {
+@Repository
+public class PostVoteRedisRepository {
 
     public static final String POST_VOTES_KEY = "post:%d:votes";
     public static final String UPVOTE_COUNT_KEY = "post:upvoteCount";
@@ -15,21 +17,21 @@ public abstract class AbstractVoteTask {
     private final RedisScript<List> postVoteScript;
     private final RedisScript<List> postUnvoteScript;
 
-    public AbstractVoteTask(
-        RedisScript<List> postUnvoteScript,
+    public PostVoteRedisRepository(
         RedisTemplate<String, String> redisTemplate,
-        RedisScript<List> postVoteScript
+        RedisScript<List> postVoteScript,
+        RedisScript<List> postUnvoteScript
     ) {
-        this.postUnvoteScript = postUnvoteScript;
         this.redisTemplate = redisTemplate;
         this.postVoteScript = postVoteScript;
+        this.postUnvoteScript = postUnvoteScript;
     }
 
     private String getPostVotesKey(Long postId) {
         return String.format(POST_VOTES_KEY, postId);
     }
 
-    public void upvote(Long postId, Long userId, VoteType voteType) {
+    public void vote(Long postId, Long userId, VoteType voteType){
         redisTemplate.execute(
             postVoteScript,
             List.of(getPostVotesKey(postId), UPVOTE_COUNT_KEY, DOWNVOTE_COUNT_KEY),
@@ -39,12 +41,11 @@ public abstract class AbstractVoteTask {
         );
     }
 
-    public void unvote(Long postId, Long userId) {
+    public void unvote(Long postId, Long userId){
         redisTemplate.execute(
             postUnvoteScript,
             List.of(getPostVotesKey(postId), UPVOTE_COUNT_KEY, DOWNVOTE_COUNT_KEY),
             postId.toString(), userId.toString()
         );
     }
-
 }
