+++ devtribe-feed-core/src/main/java/com/devtribe/global/config/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.devtribe.global.config;
+package com.devtribe.global.config.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
