+++ devtribe-feed-core/src/main/resources/scripts/post_vote.lua
