+++ devtribe-feed-core/src/main/resources/scripts/post_vote_count.lua
